Separate into smart/dumb components
