+++ source/components/MessageForm.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const MessageForm = () =>
+  <div className="messageForm">
+    Hello, world! I am a MessageForm.
+  </div>;
+
+export default MessageForm;
