+++ webpack.config.babel.js
@@ -14,7 +14,10 @@ fs.readdirSync('node_modules')
   });
 
 const sharedConfig = {
-  resolve: { root: sourcePath },
+  resolve: {
+    root: sourcePath,
+    extensions: ['', '.js', '.jsx'],
+  },
 
   output: {
     path: buildPath,
@@ -24,7 +27,7 @@ const sharedConfig = {
   module: {
     loaders: [
       { test: /\.css$/, include: [sourcePath], loader: 'style-loader!css-loader?modules&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]!postcss-loader' },
-      { test: /\.js$/, include: [sourcePath], loader: 'babel-loader' },
+      { test: /\.(js|jsx)$/, include: [sourcePath], loader: 'babel-loader' },
       { test: /\.json/, loader: 'json' },
     ],
   },
